refactor : remove CardList
