--- src/pages/CardListPage/components/CardList/CardList.style.ts
@@ -1,13 +1,0 @@
-import styled from '@emotion/styled';
-
-export const CardListContainer = styled.ul`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  margin: 0;
-  padding: 20px 0;
-  li {
-    margin: 10px 0;
-  }
-`;
