--- src/pages/CardListPage/components/CardList/index.ts
@@ -1,1 +1,0 @@
-export { default as CardList } from './CardList';
